@@ -12,27 +12,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.1.3",
+    "@angular/animations": "^5.0.0",
     "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
     "@angular/material": "^2.0.0-beta.6",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "compass": "^0.1.1",
     "compass-mixins": "^0.12.10",
     "core-js": "^2.4.1",
     "flexboxgrid": "^6.3.1",
-    "rxjs": "^5.1.0",
+    "rxjs": "^5.5.2",
+    "webpack-sources": "^1.0.2",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.9",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular/cli": "^1.5.2",
+    "@angular/compiler-cli": "^5.0.0",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
     "codelyzer": "~2.0.0",
@@ -47,6 +49,6 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.4.0"
+    "typescript": "2.4.2"
   }
 }
