@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { HttpService } from "app/services/http.service";
+// import { HttpService } from "app/services/http.service";
 import { ICertificate } from "app/models/certificate";
 import { ENDPOINTS } from "app/settings/endpoints";
 import { certificatesMock } from "app/mocks/certificates.mock";
@@ -13,7 +13,7 @@ export class CertificatesComponent implements OnInit {
 
   certificates : ICertificate[];
 
-  constructor(private httpService: HttpService) { }
+  // constructor(private httpService: HttpService) { }
 
   ngOnInit() {
     // this.httpService.get<Certificate[]>(ENDPOINTS.certificate)
